Make one deterministic call to LinkMinter
Now we're going to call the [/record endpoint](https://github.com/NYPL/link/blob/master/link-admin/src/main/java/org/nypl/repository/link_admin/LinkAdminResource.java#L187) of the link minter, not `/records`. In addition we are passing `redirect=false` which ensures our links are proxied, not redirected to.
The `uuid` parameter helps ensures a deterministically find-or-create logic.

For example:

Asking for a shortened link of "http://example.com", with a uuid of "abc-123", 4 times
will always return the same link code. If the 5th time you ask, but with a uuid of "xyz-000"
you will get a new link code.