Remove dotenv-rails
We only used dotenv for local development.
Now that local development is Dockerized there's no need for dotenv.

In the olden-days (before this commit)...running this natively/locally environment variables could be in .env.
Now however, in the variables will be sourced from the `env_file` directive in docker-compsose.yml.